requirements.txt
# Core ADK e LLM
google-adk>=0.2.0
litellm>=1.52.0
anthropic>=0.39.0

# AWS
boto3>=1.35.0
botocore>=1.35.0

# API e Web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# RAG e Embeddings
chromadb>=0.5.0
sentence-transformers>=2.3.0

# Utilidades
requests>=2.31.0
deprecated>=1.2.14
numpy<2.0.0

# Opcional mas recomendado
httpx>=0.25.0
aiofiles>=23.0.0

colorama
pandas